1-WebFrameworkBasico/requirements.txt
robotframework==4.1.2
robotframework-seleniumlibrary
datetime
dateutils
robotframework-pabot
allure-robotframework